Move IR.SelectionSet into new file
